+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/psi/impl/xquery/XQueryAnnotatedDeclPsiImpl.kt
@@ -17,13 +17,35 @@ package uk.co.reecedunn.intellij.plugin.xquery.psi.impl.xquery
 
 import com.intellij.extapi.psi.ASTWrapperPsiElement
 import com.intellij.lang.ASTNode
+import com.intellij.navigation.ItemPresentation
+import com.intellij.ui.IconManager
+import com.intellij.util.BitUtil
+import com.intellij.util.PlatformIcons
 import uk.co.reecedunn.intellij.plugin.core.sequences.children
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmAnnotation
 import uk.co.reecedunn.intellij.plugin.xpm.optree.XpmAnnotatedDeclaration
 import uk.co.reecedunn.intellij.plugin.xpm.optree.annotation
 import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryAnnotatedDecl
+import javax.swing.Icon
 
 open class XQueryAnnotatedDeclPsiImpl(node: ASTNode) : ASTWrapperPsiElement(node), XQueryAnnotatedDecl {
+    // region ElementBase
+
+    override fun getElementIcon(flags: Int): Icon? = when (this) {
+        is ItemPresentation -> {
+            val icon = getIcon(false)
+            val isLocked = BitUtil.isSet(flags, ICON_FLAG_READ_STATUS) && !isWritable
+            val baseIcon = IconManager.getInstance().createLayeredIcon(this, icon, if (isLocked) FLAGS_LOCKED else 0)
+            if (BitUtil.isSet(flags, ICON_FLAG_VISIBILITY)) when (isPublic) {
+                true -> baseIcon.setIcon(PlatformIcons.PUBLIC_ICON, 1)
+                false -> baseIcon.setIcon(PlatformIcons.PRIVATE_ICON, 1)
+            }
+            baseIcon
+        }
+        else -> null
+    }
+
+    // endregion
     // region XdmAnnotatedDeclaration
 
     override val annotations: Sequence<XdmAnnotation>